apps/enterprise/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.2
jinja2==3.1.4

